requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgihandler==0.3.9
asgiref==3.5.2
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
chardet==3.0.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
constantly==15.1.0
cryptography==38.0.4
daphne==4.0.0
distlib==0.3.6
Django==4.1.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
filelock==3.8.0
hiredis==2.0.0
hyperlink==21.0.0
idna==2.7
incremental==22.10.0
joblib==1.2.0
kombu==5.2.4
MouseInfo==0.1.3
msgpack==1.0.4
nltk==3.7
oauthlib==2.0.1
pika==1.3.1
pika-stubs==0.1.3
platformdirs==2.5.2
prompt-toolkit==3.0.32
psycopg2==2.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyJWT==2.5.0
PyMsgBox==1.0.9
pyOpenSSL==22.1.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-unsplash==1.1.0
pytweening==1.0.4
pytz==2022.4
redis==4.4.0
regex==2022.10.31
requests==2.20.0
requests-oauthlib==0.7.0
schedule==1.1.0
scheduler==0.8.0
service-identity==21.1.0
six==1.10.0
sqlparse==0.4.3
tqdm==4.64.1
Twisted==22.10.0
txaio==22.2.1
typeguard==2.13.3
typing_extensions==4.4.0
tzdata==2022.5
urllib3==1.24.3
vine==5.0.0
virtualenv==20.16.5
wcwidth==0.2.5
websockets==10.4
zope.interface==5.5.2
